+++ NEWS.md
@@ -9,8 +9,10 @@
   [#278](https://github.com/lgatto/MSnbase/issues/271)).
 - Reduce number of comparsions in in internal `fastquant_max` to get
   little speed improvents for isobaric quantification (see PR
-  [#280](https://github.com/lgatto/MSnbase/issues/280)) <2017-11-27
-  Mon>.
+  [#280](https://github.com/lgatto/MSnbase/issues/280)) <2017-11-27 Mon>.
+- MIAxE, MSnProcess and AnnotatedDataFrame coercion to list methods
+  (in related to PR [#280](https://github.com/lgatto/MSnbase/issues/284))
+  <2017-12-11 Mon>
 
 ## Changes in version 2.5.1
 - Update dependencies (see issue [#271](https://github.com/lgatto/MSnbase/issues/271))